+++ examples/arco-pro/farm.config.ts
@@ -1,60 +1,21 @@
-import { resolve } from 'node:path';
-import { defineConfig } from '@farmfe/core';
-import farmJsPluginLess from '@farmfe/js-plugin-less';
+import { resolve } from "node:path";
+import { defineConfig } from "@farmfe/core";
+import farmJsPluginLess from "@farmfe/js-plugin-less";
 
 export default defineConfig((env) => {
   return {
     compilation: {
-      input: {
-        index: './index.html'
-      },
-      sourcemap: false,
-      presetEnv: false,
-      concatenateModules: true,
-      persistentCache: false,
-      lazyCompilation: false,
       resolve: {
-        symlinks: true,
         alias: {
-          '@': resolve(process.cwd(), './src'),
-          'react-dom': resolve(process.cwd(), './node_modules/react-dom'),
-          react: resolve(process.cwd(), './node_modules/react')
-        }
+          "@": resolve(process.cwd(), "./src"),
+        },
+        dedupe: ["react", "react-dom"],
       },
-      // minify: false,
-      // mode: 'development',
-      // persistentCache: false,
-      output: {
-        path: './build',
-        filename: 'assets/[resourceName].[contentHash].[ext]',
-        assetsFilename: 'static/[resourceName].[contentHash].[ext]'
-      },
-      partialBundling: {
-        targetMinSize: 1024 * 2000,
-        groups: [
-          {
-            name: 'components',
-            test: ['src/components/.+'],
-            enforce: true,
-          },
-          {
-            name: 'xxxx',
-            test: ['src/pages/.+']
-          }
-        ]
-      },
-      progress: false
     },
     plugins: [
-      [
-        '@farmfe/plugin-react',
-        {
-          refresh: env.mode === 'development',
-          development: env.mode === 'development'
-        }
-      ],
-      '@farmfe/plugin-svgr',
+      "@farmfe/plugin-react",
+      "@farmfe/plugin-svgr",
       farmJsPluginLess(),
-    ]
+    ],
   };
 });