+++ packages/core/src/watcher/index.ts
@@ -1,6 +1,7 @@
 import EventEmitter from 'node:events';
 import { existsSync } from 'node:fs';
 import path from 'node:path';
+
 import chokidar from 'chokidar';
 import type { FSWatcher, WatchOptions } from 'chokidar';
 import glob from 'fast-glob';
@@ -19,18 +20,53 @@ import type {
 
 export const debugWatcher = createDebugger('farm:watcher');
 
-export default class Watcher {
-  private watchedFiles = new Set<string>();
+interface IWatcher {
   resolvedWatchOptions: WatchOptions;
-  watcher: FSWatcher;
   extraWatchedFiles: string[];
+  getInternalWatcher(): FSWatcher;
+  filterWatchFile(file: string, root: string): boolean;
+  getExtraWatchedFiles(compiler?: Compiler | null): string[];
+  watchExtraFiles(): void;
+  createWatcher(): Promise<void>;
+  resolveChokidarOptions(): void;
+  close(): Promise<void>;
+}
+
+export default class Watcher extends EventEmitter implements IWatcher {
+  private watchedFiles = new Set<string>();
+  public resolvedWatchOptions: WatchOptions;
+  public extraWatchedFiles: string[] = [];
+  private _watcher: FSWatcher;
 
   constructor(public config: ResolvedUserConfig) {
+    super();
     this.resolveChokidarOptions();
   }
 
+  on(
+    event: 'add' | 'addDir' | 'change' | 'unlink' | 'unlinkDir',
+    listener: (path: string) => void
+  ): this {
+    this._watcher.on(event, listener);
+    return this;
+  }
+
+  add(paths: string | ReadonlyArray<string>): this {
+    this._watcher.add(paths);
+    return this;
+  }
+
+  unwatch(paths: string | ReadonlyArray<string>): this {
+    this._watcher.unwatch(paths);
+    return this;
+  }
+
+  getWatched(): { [directory: string]: string[] } {
+    return this._watcher.getWatched();
+  }
+
   getInternalWatcher() {
-    return this.watcher;
+    return this._watcher;
   }
 
   filterWatchFile(file: string, root: string): boolean {
@@ -53,114 +89,20 @@ export default class Watcher {
   watchExtraFiles() {
     this.extraWatchedFiles.forEach((file) => {
       if (!this.watchedFiles.has(file)) {
-        this.watcher.add(file);
+        this._watcher.add(file);
         this.watchedFiles.add(file);
       }
     });
   }
 
-  async watch() {}
-
-  // async watch() {
-  //   const compiler = this.getCompiler();
-
-  //   const handlePathChange = async (path: string) => {
-  //     if (this.close) return;
-
-  //     try {
-  //       if (this.compiler instanceof NewServer && this.compiler.getCompiler()) {
-  //         await this.compiler.hmrEngine.hmrUpdate(path);
-  //       }
-
-  //       if (
-  //         this.compiler instanceof Compiler &&
-  //         this.compiler.hasModule(path)
-  //       ) {
-  //         await compilerHandler(
-  //           async () => {
-  //             const result = await compiler.update([path], true);
-  //             this.handleUpdateFinish(result, compiler);
-  //             compiler.writeResourcesToDisk();
-  //           },
-  //           this.config,
-  //           this.logger,
-  //           { clear: true }
-  //         );
-  //       }
-  //     } catch (error) {
-  //       this.logger.error(error);
-  //     }
-  //   };
-
-  //   const filesToWatch = [this.config.root, ...this.getExtraWatchedFiles()];
-  //   this.watchedFiles = new Set(filesToWatch);
-  //   this.watcher ??= createWatcher(this.config, filesToWatch);
-
-  //   this.watcher.on('change', (path) => {
-  //     if (this.close) return;
-  //     handlePathChange(path);
-  //   });
-
-  //   if (this.compiler instanceof NewServer) {
-  //     this.compiler.hmrEngine?.onUpdateFinish((result) =>
-  //       this.handleUpdateFinish(result, compiler)
-  //     );
-  //   }
-  // }
-
-  // async watchConfigs(callback: (files: string[]) => void) {
-  //   const filesToWatch = Array.from([
-  //     ...(this.config.envFiles ?? []),
-  //     ...(this.config.configFileDependencies ?? []),
-  //     ...(this.config.configFilePath ? [this.config.configFilePath] : [])
-  //   ]).filter((file) => file && existsSync(file));
-  //   const chokidarOptions = {
-  //     awaitWriteFinish:
-  //       process.platform === 'linux'
-  //         ? undefined
-  //         : {
-  //             stabilityThreshold: 10,
-  //             pollInterval: 80
-  //           }
-  //   };
-  //   this.watcher ??= createWatcher(this.config, filesToWatch, chokidarOptions);
-
-  //   this.watcher.on('change', (path) => {
-  //     if (this.close) return;
-  //     if (filesToWatch.includes(path)) {
-  //       callback([path]);
-  //     }
-  //   });
-  //   return this;
-  // }
-
-  // private handleUpdateFinish(updateResult: JsUpdateResult, compiler: Compiler) {
-  //   const addedFiles = [
-  //     ...updateResult.added,
-  //     ...updateResult.extraWatchResult.add
-  //   ].map((addedModule) =>
-  //     compiler.transformModulePath(this.config.root, addedModule)
-  //   );
-
-  //   const filteredAdded = addedFiles.filter((file) =>
-  //     this.filterWatchFile(file, this.config.root)
-  //   );
-
-  //   if (filteredAdded.length > 0) {
-  //     this.watcher.add(filteredAdded);
-  //   }
-  // }
-
   async createWatcher() {
     const compiler = await createInlineCompiler(this.config, {
       progress: false
     });
-    // TODO type error here
-    // @ts-ignore
     const enabledWatcher = this.config.watch !== null;
     const files = [this.config.root, ...this.getExtraWatchedFiles(compiler)];
 
-    this.watcher = enabledWatcher
+    this._watcher = enabledWatcher
       ? (chokidar.watch(files, this.resolvedWatchOptions) as FSWatcher)
       : new NoopWatcher(this.resolvedWatchOptions);
   }
@@ -199,10 +141,10 @@ export default class Watcher {
   }
 
   async close() {
-    if (this.watcher) {
+    if (this._watcher) {
       debugWatcher?.('close watcher');
-      await this.watcher.close();
-      this.watcher = null;
+      await this._watcher.close();
+      this._watcher = null;
     }
   }
 }
@@ -244,7 +186,7 @@ export async function handlerWatcher(
   const logger = resolvedUserConfig.logger;
   const watcher = new Watcher(resolvedUserConfig);
   await watcher.createWatcher();
-  watcher.watcher.on('change', async (file: string | string[] | any) => {
+  watcher.on('change', async (file: string | string[] | any) => {
     file = normalizePath(file);
     // TODO restart with node side v2.0 we may be think about this feature
     // const shortFile = getShortName(file, resolvedUserConfig.root);
@@ -277,7 +219,7 @@ export async function handlerWatcher(
       );
 
       if (filteredAdded.length > 0) {
-        watcher.watcher.add(filteredAdded);
+        watcher.add(filteredAdded);
       }
     };