+++ rust-plugins/replace-dirname/src/lib.rs
@@ -7,7 +7,7 @@ use farmfe_core::{
   error::CompilationError,
   plugin::{Plugin, PluginProcessModuleHookParam},
   swc_common::DUMMY_SP,
-  swc_ecma_ast::{self, Expr, Lit, MemberExpr, MemberProp, Module, Str},
+  swc_ecma_ast::{self, Callee, Expr, Lit, MemberExpr, MemberProp, MetaPropKind, Module, Str},
 };
 use std::{env, path::Path, sync::Arc};
 use url::Url;
@@ -100,9 +100,22 @@ pub fn replace_dirname_with_ast(ast: &mut Module, dir_path: &str, file_path: &st
           }
           _ => {}
         },
+
+        Expr::Call(call_expr) => {
+          for arg in &mut call_expr.args {
+            self.visit_mut_expr(&mut arg.expr);
+          }
+
+          if let Callee::Expr(expr) = &mut call_expr.callee {
+            self.visit_mut_expr(expr);
+          }
+        }
+
         Expr::Member(MemberExpr { obj, prop, .. }) => {
+          // #2062
+          self.visit_mut_expr(obj);
           if let Expr::MetaProp(meta_prop) = &**obj {
-            if meta_prop.kind == swc_ecma_ast::MetaPropKind::ImportMeta {
+            if meta_prop.kind == MetaPropKind::ImportMeta {
               if let MemberProp::Ident(ident) = &prop {
                 if ident.sym == "url" {
                   if let Ok(file_path) = Url::from_file_path(self.file_path) {